Move mismanaged enum
