+++ static/app/views/starfish/views/webServiceView/starfishView.tsx
@@ -15,7 +15,7 @@ import {useTheme} from '@emotion/react';
 import {getInterval} from 'sentry/components/charts/utils';
 import {t} from 'sentry/locale';
 import {tooltipFormatterUsingAggregateOutputType} from 'sentry/utils/discover/charts';
-import {RateUnits} from 'sentry/utils/discover/fieldRenderers';
+import {RateUnits} from 'sentry/utils/discover/fields';
 import {DiscoverDatasets} from 'sentry/utils/discover/types';
 import {MutableSearch} from 'sentry/utils/tokenizeSearch';
 import usePageFilters from 'sentry/utils/usePageFilters';
